reorganized tests, created adCard and adListFilter tests
