+++ frontend/src/__test__/pages/AdList/AdListFilter/AdListFilter.test.jsx
@@ -4,28 +4,22 @@ import { Provider } from 'react-redux'
 import { render, screen } from '@testing-library/react'
 import { BrowserRouter } from 'react-router-dom'
 import { describe, it, expect } from 'vitest'
-import AdList from '../../pages/AdList/AdList'
-import store from '../../store/store'
+import AdListFilter from '../../../../pages/AdList/AdListFilter/AdListFilter'
+import store from '../../../../store/store'
 
 axios.defaults.baseURL = import.meta.env.VITE_API_URL
 
-describe('AdList', () => {
-  beforeEach(() => {
-    window.localStorage.clear()
-  })
-
-  it('should render the AdList', () => {
-    window.localStorage.setItem('token', 'true')
-    window.localStorage.setItem('refreshToken', 'false')
+describe('AdListFilter', () => {
+  it('should render the AdListFilter', () => {
     render(
       <BrowserRouter>
         <Provider store={store}>
-          <AdList />
+          <AdListFilter />
         </Provider>
       </BrowserRouter>
     )
 
-    const adListStyled = screen.getByTestId('adListStyled')
+    const adListStyled = screen.getByText('Filtros')
     expect(adListStyled).toBeInTheDocument()
   })
 })
